requirements.txt
rich>=13.7
markdown>=3.5
pyparsing>=3.0.0
python-dotenv>=1.0.1
instructor>=0.4.8
PyYAML>=6.0
praisonaiagents>=0.0.4
duckduckgo_search
praisonai_tools
openai>=1.12.0
langgraph>=0.3.25
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.2
pytest>=7.0.0
